requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
passlib==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
pydantic==2.4.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
certifi==2023.11.17